Add the service package

+++ src/main/java/br/com/possenti/repository/GoalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.possenti.model.Goal;
 
-public interface GoalRepository extends CrudRepository<Goal, Long>{
+public interface GoalRepository extends CrudRepository<Goal, Integer>{
 
 }
